base-station: color the mission path by expected comms quality
Even with the radio arc and mobile-coverage rings drawn, operators
still had to eyeball whether each leg of a planned mission would
stay covered — surveys in particular cross ring borders many times
and made coverage gaps easy to miss.

- Sample each mission segment and classify the midpoint against
  the radio arc range (with tracking-antenna behavior, so the
  whole arc counts as in-range) and the nearest mobile-coverage
  ring border, in `libs/baseStation/missionPathSignal.ts`.
- Render the mission and survey paths colored Good/Marginal/Bad
  using the new `useMissionPathSignal` composable, merging
  contiguous same-risk samples to keep the line clean.
- Extract the shared OpenCellID/Overpass helpers into
  `libs/baseStation/mobileCoverage.ts` so both the overlay and the
  mission sampler reuse the same logic.
- Add a "Show comms coverage on path" toggle to the mission
  estimates panel, gated by base-station availability.
