+++ src/views/MissionPlanningView.vue
@@ -619,7 +619,7 @@
   </SideConfigPanel>
   <HomePositionSettingHelp v-model="showHomePositionNotSetDialog" />
   <PoiManager ref="poiManagerRef" />
-  <BaseStationConfigPanel />
+  <BaseStationConfigPanel is-mission-planning-context />
   <BaseStationContextPopup />
   <PoiMapArrows
     :map-ready="mapReady"
@@ -627,8 +627,11 @@
     :show-poi-arrows="true"
     :show-home-arrow="true"
     :show-vehicle-arrow="true"
+    :show-base-station-arrow="baseStationStore.config.enabled"
     :vehicle-position="vehiclePosition"
     :home="home"
+    :base-station="baseStationStore.config.enabled ? baseStationStore.config.position ?? undefined : undefined"
+    :base-station-color="baseStationStore.config.coverageColor"
     :map-center="mapCenter"
     :zoom="zoom"
     :target-follower="targetFollower"
@@ -695,6 +698,7 @@ import RadialMenu, { type RadialMenuItem } from '@/components/RadialMenu.vue'
 import SideConfigPanel from '@/components/SideConfigPanel.vue'
 import { useBaseStation } from '@/composables/baseStation/useBaseStation'
 import { useBaseStationOverlay } from '@/composables/baseStation/useBaseStationOverlay'
+import { useMissionPathSignal } from '@/composables/baseStation/useMissionPathSignal'
 import { useInteractionDialog } from '@/composables/interactionDialog'
 import { provideMapContext } from '@/composables/map/useMapContext'
 import { useSnackbar } from '@/composables/snackbar'
@@ -705,6 +709,7 @@ import {
   useMissionEstimates,
 } from '@/composables/useMissionEstimates'
 import { useOfflineTiles } from '@/composables/useOfflineTiles'
+import { buildMissionPathDisplaySegments, MISSION_COVERAGE_RISK_COLORS } from '@/libs/baseStation/missionPathSignal'
 import { MavType } from '@/libs/connection/m2r/messages/mavlink2rest-enum'
 import { MavCmd } from '@/libs/connection/m2r/messages/mavlink2rest-enum'
 import type { NoiseTileOptions } from '@/libs/map/map-tile-fallback'
@@ -744,6 +749,8 @@ const interfaceStore = useAppInterfaceStore()
 const widgetStore = useWidgetManagerStore()
 const baseStationStore = useBaseStation()
 const missionEstimates = useMissionEstimates()
+const { isPathSignalAvailable: isMissionPathSignalAvailable, mobileCoverageCircles: missionMobileCoverageCircles } =
+  useMissionPathSignal()
 const { height: windowHeight } = useWindowSize()
 
 const { showDialog, closeDialog } = useInteractionDialog()
@@ -1292,6 +1299,7 @@ const clearCurrentMission = (): void => {
     planningMap.value?.removeLayer(missionWaypointsPolyline.value)
     missionWaypointsPolyline.value = null
   }
+  removeMissionPathSignalLayer()
   clearSurveyPath()
   selectedSurveyId.value = ''
   lastSelectedSurveyId.value = ''
@@ -2793,6 +2801,7 @@ const clearSurveyPath = (): void => {
   surveyEdgeAddMarkers.forEach((marker) => marker.remove())
   surveyPolygonVertexesMarkers.value = []
   surveyPolygonVertexesPositions.value = []
+  renderMissionPathSignal()
 }
 
 watch([isCreatingSurvey, isCreatingSimplePath], (isCreatingNow) => {
@@ -2875,6 +2884,7 @@ const checkAndRemoveSurveyPath = (): void => {
   surveyPathLayer.value = null
   surveyTurnaroundLayers.value.forEach((layer) => planningMap.value?.removeLayer(layer as unknown as L.Layer))
   surveyTurnaroundLayers.value = []
+  renderMissionPathSignal()
 }
 
 const createSurveyPath = (): void => {
@@ -2925,6 +2935,8 @@ const createSurveyPath = (): void => {
         }).addTo(toRaw(planningMap.value)!)
       )
     }
+
+    renderMissionPathSignal()
   } catch (error) {
     showDialog({
       variant: 'error',
@@ -3970,6 +3982,8 @@ onUnmounted(() => {
   stopTileFallbackWatcher?.()
   stopTileFallbackWatcher = undefined
 
+  removeMissionPathSignalLayer()
+
   // Reset the map context so descendants stop reacting to the destroyed instance
   mapContext.mapReady.value = false
   mapContext.map.value = undefined
@@ -4122,9 +4136,89 @@ watch(zoom, () => {
 })
 
 const missionWaypointsPolyline = shallowRef<L.Polyline | null>(null)
+const missionPathSignalLayer = shallowRef<L.LayerGroup | null>(null)
+const waypointPolylineBaseStyle: L.PolylineOptions = { opacity: 1, color: '#3388ff', weight: 3 }
+const surveyPathLayerBaseStyle: L.PolylineOptions = { color: '#2563EB', weight: 3, opacity: 0.8 }
 
 const getMissionPathLatLngs = (): L.LatLng[] =>
   missionStore.currentPlanningWaypoints.map((waypoint) => L.latLng(waypoint.coordinates[0], waypoint.coordinates[1]))
+
+const flattenPolylineCoordinates = (polyline: L.Polyline): WaypointCoordinates[] => {
+  const latlngs = polyline.getLatLngs()
+  if (!latlngs.length) return []
+  const flat = latlngs[0] instanceof L.LatLng ? (latlngs as L.LatLng[]) : (latlngs as L.LatLng[][]).flat()
+  return flat.map((latlng) => [latlng.lat, latlng.lng] as WaypointCoordinates)
+}
+
+const getMissionCoveragePathCoordinates = (): WaypointCoordinates[] => {
+  if (surveyPathLayer.value) {
+    return flattenPolylineCoordinates(surveyPathLayer.value)
+  }
+  return missionStore.currentPlanningWaypoints.map((waypoint) => waypoint.coordinates)
+}
+
+const restoreMissionPathLineStyles = (): void => {
+  if (missionWaypointsPolyline.value) {
+    missionWaypointsPolyline.value.setStyle(waypointPolylineBaseStyle)
+  }
+  if (surveyPathLayer.value) {
+    surveyPathLayer.value.setStyle(surveyPathLayerBaseStyle)
+  }
+}
+
+const removeMissionPathSignalLayer = (): void => {
+  if (!missionPathSignalLayer.value) return
+  missionPathSignalLayer.value.clearLayers()
+  planningMap.value?.removeLayer(missionPathSignalLayer.value)
+  missionPathSignalLayer.value = null
+}
+
+const renderMissionPathSignalImmediate = (): void => {
+  if (!planningMap.value) return
+  removeMissionPathSignalLayer()
+
+  if (!missionStore.showMissionPathSignalStrength || !isMissionPathSignalAvailable.value) {
+    restoreMissionPathLineStyles()
+    return
+  }
+
+  const pathCoordinates = getMissionCoveragePathCoordinates()
+  if (pathCoordinates.length < 2) {
+    restoreMissionPathLineStyles()
+    return
+  }
+
+  const isSurveyPath = !!surveyPathLayer.value
+  const displaySegments = buildMissionPathDisplaySegments(
+    baseStationStore.config,
+    pathCoordinates,
+    missionMobileCoverageCircles.value
+  )
+
+  if (missionWaypointsPolyline.value) {
+    missionWaypointsPolyline.value.setStyle({ opacity: 0 })
+  }
+  if (surveyPathLayer.value) {
+    surveyPathLayer.value.setStyle({ opacity: 0 })
+  }
+
+  missionPathSignalLayer.value = L.layerGroup()
+  for (const segment of displaySegments) {
+    const polyline = L.polyline([segment.start, segment.end], {
+      color: MISSION_COVERAGE_RISK_COLORS[segment.risk],
+      weight: 3,
+      opacity: isSurveyPath ? 0.8 : 1,
+      interactive: false,
+      ...(isSurveyPath ? { className: 'survey-path' } : {}),
+    })
+    polyline.addTo(missionPathSignalLayer.value)
+  }
+  missionPathSignalLayer.value.addTo(planningMap.value)
+}
+
+// Coalesces redraws caused by survey edits, waypoint drags, and config changes within the
+// same animation frame; long enough to flatten typed-input bursts, short enough to feel live.
+const renderMissionPathSignal = useDebounceFn(renderMissionPathSignalImmediate, 100)
 
 watch(
   () => missionStore.currentPlanningWaypoints.map((waypoint) => waypoint.coordinates.slice()),
@@ -4143,8 +4237,31 @@ watch(
     } else {
       missionWaypointsPolyline.value.setLatLngs(missionPathLatLngs)
     }
+
+    renderMissionPathSignal()
   },
   { immediate: true, deep: true }
+)
+
+watch(
+  () => [
+    isMissionPathSignalAvailable.value,
+    missionMobileCoverageCircles.value,
+    missionStore.showMissionPathSignalStrength,
+    baseStationStore.config.enabled,
+    baseStationStore.config.position,
+    baseStationStore.config.commsType,
+    baseStationStore.config.antenna,
+    baseStationStore.config.baseStationAntennaHeightMeters,
+    baseStationStore.config.vehicleHasBlueBoatAntennaMast,
+    baseStationStore.config.mobileCoverage.provider,
+    baseStationStore.config.mobileCoverage.openCellIdOperator,
+    baseStationStore.config.mobileCoverage.osmOperator,
+  ],
+  () => {
+    renderMissionPathSignal()
+  },
+  { deep: true }
 )
 
 // Create polyline for the vehicle path using a dedicated Canvas renderer to prevent performance issues
